pyproject.toml
[project]
name = "elixir-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "berkeleydb>=18.1.14",
    "gitpython>=3.1.44",
    "mcp>=1.6.0",
]
